vim-patch:9.1.1483: not possible to translation position in buffer (#37099)
Problem:  not possible to translation position in buffer
Solution: use _() macro to mark the output as translatable
          (Emir SARI)

Row/Column indicator separator is currently not customizable. Some
languages have a space after the comma as the usual practice, plus this
would help translators use a custom separator like colons if necessary.

Additionally, after a save, the line and the byte indicator is also
hardcoded, this enables i18n for that as well.

closes: vim/vim#17608

https://github.com/vim/vim/commit/81f9815831c7b3fc9ea1e0164e27e12f321ff2c3

Co-authored-by: Emir SARI <emir_sari@icloud.com>